[REFACTOR] 서재 평가 삭제 로직 분리 및 Application으로 이전
